wandb/requirements.txt
ipykernel==6.29.4
ipython==8.23.0
numpy==1.26.4
pandas==2.2.2
python-dotenv==1.0.1
scikit-learn==1.4.2
wandb==0.16.6